Refactor: QuestService Interface를 상속받는 구조로 변경
